Fixed two bugs due to refactoring
